Slice the bot to two parts: Bot and Client
Bot is the UI to users, that will use buttons and Client is the backend
that is subscribed to channels. Actually, second one is userbot. This
is done because bots can't join channels, which is critical for us.

For now, there is only a few moments that need to be done before this
bot can be used by everyone. One of those are logging, Docker support
and good README. I am planning to add them shortly.